+++ src/main/java/com/rewedigital/composer/ComposerApplication.java
@@ -6,12 +6,12 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.function.BiFunction;
 
+import com.rewedigital.composer.application.ProxyHeaderMiddleware;
 import com.rewedigital.composer.application.RequestHandler;
 import com.rewedigital.composer.application.RequestHandlerFactory;
 import com.rewedigital.composer.caching.HttpCacheModule;
 import com.rewedigital.composer.client.ErrorHandlingClientDecoratingModule;
 import com.rewedigital.composer.client.WithIncomingHeadersClientDecoratingModule;
-import com.rewedigital.composer.proxy.ProxyHeaderMiddleware;
 import com.spotify.apollo.AppInit;
 import com.spotify.apollo.Environment;
 import com.spotify.apollo.core.Service;
