+++ src/test/java/com/rewedigital/composer/application/ComposingRequestHandlerTest.java
@@ -1,4 +1,4 @@
-package com.rewedigital.composer.proxy;
+package com.rewedigital.composer.application;
 
 import static com.rewedigital.composer.routing.RouteTypeName.PROXY;
 import static com.rewedigital.composer.routing.RouteTypeName.TEMPLATE;
